Beginning self-hosting
This will likely be a gigantic undertaking which will probably take a
very long time to complete. In order to gauge progress, I've created a
testing strategy to maintain parity with the existing rust
implementation. So far, there is a Lexer which outputs tokens, and the
only token it can do is an integer. LONG way to go!