Add tokenization for hex integers
Selfhost: tokenize hexadecimal integer literals as Int tokens. However,
this highlighted an interesting issue with the way garbage collection
works with respect to enums (I think, I'm not 100% sure that this was
the issue). What I _think_ is the case is that, since tagged union enum
variants are implemented as a pointer whose upper bits are set to be the
variant type, the garbage collector isn't able to mark the allocated
objects and it frees them prematurely.