Fixing enums w.r.t. GC
Enums had been broken ever since I integrated the bdwgc garbage
collector into the runtime. This is because of the way I "cleverly"
implemented them as being 64-bits which were either an integer for
constant variants (just the runtime typeid for comparisons) or a
pointer whose upper bits represented the runtime typeid. Because of the
fact that the pointer value needed to be masked in order to be
recognizable, the garbage collector did not properly mark it and the
memory was prematurely freed, leading to weird issues in the selfhosting
compiler when attempting to test the lexer. This _should_ be all
resolved now.

In addition to this, I also cleaned up the trait value encoding/decoding
logic (with even more room for improvement in the future!).