Introduce LexerError into the selfhosted Lexer
Selfhost: Use Result to return an error when an unexpected token (or the
end of the file) is discovered while parsing hex/binary integer
literals. Also add proper printing/formatting of the LexerError message,
and add scaffolding to the testing "framework" for validating it against
the rust implementation.
Prelude: Add `String#split`/`String#lines` to aid in printing the
LexerError message in the selfhosted Lexer.
Lexer: The original lexer actually didn't care about invalid hex/binary
literals following the `0x`/`0b`! I felt like this was a miss, and it's
something that I wanted to include in the selfhosted lexer so I added it
to the reference implementation (especially because the way testing
works, it compares the new output vs the old, so the old output kinda
_has_ to be "correct"...).
Typechecker2: I had to modify how tagged union enum variants were
typechecked in order to allow for the selfhosted lexer to print
LexerErrors correctly in `LexerError#getMessage`. Previously, tagged
union enum variants weren't fully populated until pass 2 (in prior
passes they were just stubbed out with a bogus `FuncId`). However, this
made it impossible to destructure the variant in a match statement
within an enum's instance method body, since at the time we're
typechecking instance methods, the bogus `FuncId` hasn't yet been
replaced with a valid `FuncId`. So, I do a similar thing as we do with
functions in other places, and during enums pass 1 I do a typechecking
of the tagged union enum variant and allow for partial completion; I
later come back and patch any incomplete parameters in enums pass 2.
LLVM2: When generating toString/hash logic for generic enum variants, if
the generic argument is itself generic, it failed to resolve the latter
generic during toString-generation. To fix this, I need to do this kinda
weird thing where I extend the resolved_generics multiple times. I'm not
sure if this scales well (for example, what about if I were to do
`Result<Result<Int, Int>[], Int>`?). But I'm sure it's good enough for
now.