Selfhosted parser: parsing literals
Set up the scaffolding for the selfhosted implementation of the Parser,
including a test suite which compares the output (which is the AST
converted into a JSON representation) to a file. I chose not to compare
against the rust implementation because it's likely that there will be
deviations (and implementing a JSON conversion for the nodes in the rust
implementation would be super cumbersome).