Selfhost: Assignment
In the selfhosted rewrite, I made the decision to change assignment
_expressions_ to be _statements_ instead. The flexibility of using
assignments as expressions (eg. `val x = (a = 1)`; both x and a would be
`1`) is nice, but it offers more confusion that it's worth, especially
when it comes time to generate LLVM code. This does represent a
divergence from the reference implementation.