Selfhost: parsing decorators
Add parsing of decorators to the selfhosted parser. This also adds
support for specially-designated parse-time decorators which allow
for function bodies to be empty (`Stub`, `Intrinsic`, `CBinding`). I
imagine that eventually the `Stub` decorator will become unnecessary as
it's a holdover from the partially-implemented `prelude.abra` module.
Once all functions are implemented in that module, I can't see a reason
why `Stub` would be necessary anymore.