Selfhost: parsing exports
Add parsing for exports to the selfhosted parser. This supports
exporting val/var/func at the moment (though type/enum will come in the
future when those are actually implemented). Also in this changeset, I
added decorators to val/var since I realized I had forgotten to do that
in the previous changeset.