Selfhost: Minor cleanup - `_expectNextTokenKind`
Create helper method which consumes a token, and asserts its type,
returning an error if the token type does not match and the token itself
if if does. It's worthwhile to note that some of the TokenKind variants
have associated data - these values will NOT be compared when asserting
TokenKind equality (and when passing in a desired TokenKind to the
method, such values are also disregarded).