Selfhost: begin work on typechecker
This changeset lays out the scaffolding for the selfhosted typechecker.
It mimics the ModuleLoader/Project structure of the rust-based reference
implementation, although of course it's a lot simpler because I don't
have to worry about managing lifetimes. It's also kinda nice to "start
over" in a way, since a lot of the stuff I had done regarding the
competing `ModuleId` types between the parser and typechecker2 in the
reference implementation made things a bit muddy, as well as the
confusing uses of paths to denote a module. I'm hoping to simplify
things this time around.
Since the typechecker is the thing that loads in the file, and tokenizes
and parses it, the `TypecheckerError` is a combination of several
different types of errors. The first typechecker test files verify that
errors surfaced by the Lexer and Parser bubble up from the Typechecker.