Trying again to fix the weird test case issue where the github test r… (#402)
I've rewritten all of these commit messages because they were basically all "argh please work". This issue was particularly annoying to fix because for whatever reason I was unable to reproduce it locally, so I kept needing to push up incremental debugging messages for the github action runner (ubuntu) to run.

The issue here was that I was passing a `String#_buffer` field (which is basically just a `char*` under the hood) into `libc.open` as the file path in `fs::readFile`. This is a problem because `String`s in abra were previously _not_ NULL-terminated, because there had never been a need; for all string operations, the `#length` field can be used. However, since I was passing the raw `_buffer` into a C function (which expects NULL-terminated strings), I was implicitly passing in additional garbage that happened to be adjacent in memory until a `\0` happened to be encountered. This resulted in the file being unable to be read (because the filename I was "providing" of course didn't exist). However, this was difficult to figure out because whenever I print `String`s in abra,  just use the known `#length`.

To solve this issue, I could have added some additional logic to surround passing `String` instances into C functions. This could have been like
```
func cstring(str: String): Pointer<Byte> {
  val buf: Pointer<Byte> = Pointer.malloc(str.length + 1) // alloc extra byte for \0
  buf.copyFrom(str._buffer, str.length)
  return buf
}
```
However, that's pretty easy to forget in my opinion, and it's also wasteful because it copies the entire string. Instead, I just modified the `String.withLength` method to allocate an extra byte to hold the `\0` byte. This is unused in most cases (because I'll still use `String#length` whenever doing operations regarding the string's length), but I think is in general a sane contract, especially in low-level places where `libc` functions are frequently called.