Selfhost: refactor option types
Rather than modeling Option types (`T?`) internally as a special-case
generic struct, just have a `nullable` flag on the `Type` instance which
tracks whether `None` is an acceptable value for that type. This makes a
lot of things a lot easier when it comes to type comparisons, as generic
substitutions are no longer required. An example like `[1, None, 3]` in
the original reference implementation required "type unification" in
order to work (although that also handled other cases as well); this new
approach simply recognizes that the array contains _nullable_ versions
of the inner type and proceeds accordingly (it also goes back and marks
the items' type as `nullable = true` too). This will be especially
helpful when it comes time to generate LLVM output, since Option
instances need to be strongly-typed.