Selfhost: typechecking array literals (#405)
I think some changes are due here for how Option types are represented
internally. Representing them as a pseudo-generic instance makes a lot
of the logic weird and difficult to reason about, since this is a
special-case instance with lots of additional generic-resolving logic
built in. Instead, I think having an `#isNullable` field on the `Type`
instance itself will allow _much_ clearer expression of intent.