Selfhost: typechecking if expr/stmt (#407)
Add typechecking of if-expressions/-statements to the selfhosted
typechecker. There's a good change that this will need to be amended as
the work goes on - there's a lot of edge cases that come up
(historically) when trying to differentiate between an expression and
a statement, and a lot of other edge cases wher the type of the
else-block helps inform the type of the then-block. For example:
```
val _ = if true { [] } else { [1, 2, 3] }
// or
val _ = if true None else 123
```

There's also the case too of needing to represent nullable types; so
this example should be typechecked as follows:
```
val _ = if true {
  if true  // This expression becomes nullable: `Int?`
    123      // <- this remains an `Int`
  else
    456      // <- this remains an `Int` as well
} else {
  None     // <- this lifts the expression up into nullable-space
}
```
The type of the _nested_ if-expression in the then-block becomes wrapped
as a nullable, but not the expressions of its then-/else-blocks.