Selfhost: typechecking function declarations
This contains passes 1 and 2 of function declaration typechecking. This
is proceeding in a similar way as the reference implementation, but
maybe a bit simpler (it could also be the case that I'm missing
functionality and I haven't realized yet...). There's still a whole
bunch of logic that needs to be implemented to properly handle default
parameter values, such as when those expressions contain function calls
themselves. However, this will come later since invocation has yet to be
covered.