Selfhost: typechecking function declarations (#408)
* Selfhost: typechecking function declarations

This contains passes 1 and 2 of function declaration typechecking. This
is proceeding in a similar way as the reference implementation, but
maybe a bit simpler (it could also be the case that I'm missing
functionality and I haven't realized yet...). There's still a whole
bunch of logic that needs to be implemented to properly handle default
parameter values, such as when those expressions contain function calls
themselves. However, this will come later since invocation has yet to be
covered.

* Add 'unreachable' helper to typechecker

* Child scopes (for if and function)

* Finishing up pass 3 for function declarations

Pass 3 takes care of typechecking the function's body nodes, and is the
last step in typechecking a function declaration. By the time this
pass is run, all other declarations will have effectively been
pre-declared and should be accessible and fully-typed.

There's still the issue of default-valued parameters having expressions
which reference other functions; I haven't tackled that yet.