Selfhost: typechecking return statements
This adds typechecking support for `return` statements to the selfhosted
typechecker. This also introduces the concept of the `Never` type, which
is the type of a `return` statement when used as an expression (like as
the last value in an if-expression's then/else block). This type will
also be used to represent functions which end the program (like
`todo()`, `unreachable()`, `exit()`, etc).

In addition to `return` statements, this also scaffolds out the concept
of other scope terminators (`break`, and `continue`), although their
implementations are forthcoming. But basically, any code that follows a
previous expression/statement in which all branches have a terminator is
not accessible and results in an error.