Selfhost: typechecking invocation of functions
This is the first portion of the work to typecheck function invocation.
Once methods are implemented, I will expand upon this work, as well as
invocation of arbitrary expressions as functions.

This also adds two new methods to the `Array` type in the `prelude`
module: `Array#keyBy` and `Array#indexBy`. I only use `#keyBy` in the
selfhosted typechecker, but I thought it made sense to add `#indexBy`
while I was at it.