Selfhost: typechecking generic functions (#414)
(This also adds the `Map` type, just as a way of testing generic
resolutions for multiple type arguments)

Typechecking function declaration as well as invocation of functions
with type parameters. I'm almost certain that I missed stuff here
because of the difficulties I had originally in the reference
implementation, but time will tell I'm sure. This also takes an
interesting approach which I did not do the first time around, where a
parameter that would fail to typecheck due to unfilled holes in the
resulting typed expression's type get "re-enqueued" to try again after
the rest of the parameters have been typechecked. The idea here is that
the rest of the parameters should provide enough of the type information
in order to fill in the holes, so that when the original parameter is
tried again, the holes can be filled. If not, then we're likely in a
spot where the holes cannot be filled without explicit type annotations
(or explicitly passed type arguments, which is also implemented in this
changeset).