Selfhost: typechecking invocation as param default value (#415)
Previously I had skipped over this part due to its complexity; if
functions have default-valued parameters which reference other
incompletely-typed functions (or even the same function, recursively),
we'd have incomplete type information. This includes default-value
expressions which have an invocation nested deeply in the AST. To get
around this, in pass 2 keep track of the parameters which need
"revisiting"; these parameters will be re-processed at the start of pass
3, right before typechecking the function's body. To determine whether a
parameter needs to be "revisisted", we track if it contains _any_
function invocation (even if that function invocation would have
resulted in a successful typecheck, it's just easier this way), or if it
references any variables (this includes prior function arguments, which
may or may not have incomplete default-value expressions). If any of
these are true, then the parameter needs to be revisited in pass 3. The
variable referencing case will also become necessary in cases where
globals and/or other (captured) variables are referenced in any of the
parameter's default-value expressions; I will get to this when I
implement closure tracking.