Selfhost: typechecking binary assignment ops (#425)
When parsing binary assignment operators (eg. `+=`), emit a modified AST
structure. This saves typechecking from having to re-do all of the
LHS validations that are necessary for Assignments, and also cleans up
the typechecking logic for Binary nodes as well.

This also fixes a bug where accessing a field of a type which happened
to be nullable (eg. `type Foo { a: Int? }`) would not be correctly typed
(`Foo(a: 123).a` would be of type `Int` rather than `Int?`).