Selfhost: typechecking imports
I started working on typechecking imports, starting with relative
imports. There are a couple of error cases handled so far, with support
for circular dependency detection. The error reporting for circular
dependency detection however is subpar; it ought to surface the error on
the original import which started the cycle, rather than the current
behavior where the error is surfaced on the last module prior to the
cycle repeating.