Selfhost: typechecking imports (#427)
* Selfhost: typechecking imports

I started working on typechecking imports, starting with relative
imports. There are a couple of error cases handled so far, with support
for circular dependency detection. The error reporting for circular
dependency detection however is subpar; it ought to surface the error on
the original import which started the cycle, rather than the current
behavior where the error is surfaced on the last module prior to the
cycle repeating.

* Don't add the typed module to the project twice

* Adding test cases for circular deps

* Adding test_dir 'macro' to test

* Typechecking variable imports

* Sort modules in selfhosted typechecker json

Implement dual-pivot quicksort in `Array#sortBy` to enable sorting
modules by id in the json output for multi-module typechecker tests.

* Typechecking function imports/exports

* Importing/exporting types and enums