Making incremental progress
Getting further and further along with typechecking the prelude module
using the selfhosted typechecker. Getting generics working properly for
method calls (especially in cases where the generic value is known to be
a _generic_ in scope), filling in the prelude structs and using those
interchangeably with their `TypeKind.Primitive___` counterparts.