Selfhost: typechecking prelude (#430)
* Selfhost: typechecking prelude

Working towards typechecking the existing prelude module prior to
typechecking the entrypoint module. There's quite a few small things
that need to happen in order for this to work, and this isn't yet
working fully.

* Making incremental progress

Getting further and further along with typechecking the prelude module
using the selfhosted typechecker. Getting generics working properly for
method calls (especially in cases where the generic value is known to be
a _generic_ in scope), filling in the prelude structs and using those
interchangeably with their `TypeKind.Primitive___` counterparts.

* Updating tests to reflect additional modules

* Removing (gigantic) accidentally-committed prelude.out.json file

* Minor cleanups