Selfhost: typechecking lambdas
For the most part, lambdas are working correctly. I was able to leverage
a lot of the work I had previously done in typechecking function
parameters and bodies, with minimal changes to make it work for both.
There still is the remaining task of getting generics to work (see the
comment in `Set#forEach`).