Selfhost: typechecking lambdas (#432)
* Selfhost: typechecking lambdas

For the most part, lambdas are working correctly. I was able to leverage
a lot of the work I had previously done in typechecking function
parameters and bodies, with minimal changes to make it work for both.
There still is the remaining task of getting generics to work (see the
comment in `Set#forEach`).

* Ensuring known generics are always used

When typechecking lambdas whose parameters (or return values) are "known
generics" (ie, type parameters to the current function, as well as the
type, if the function is an instance method), allow those types to
represent valid types for the value (rather than the prior behavior of
treating an unresolved generic as an unfilled hole).