Option values are no longer implicit
In the reference (rust) implementation of the v2 typechecker/compiler,
as well as in the WIP selfhosted compiler, Option/nullable (since I
often used them interchangeably) types are _no longer_ implicit. In the
past, they were implicitly inferred, and a type could become upgraded to
be nullable if an instance of `None` were used in the same type
resolution; for example:
```
val a = if true 123 else None
```
The type of `a` here is `Int?`. However, this makes a lot of things in
the type system a _lot_ more difficult, and results in a lot more
decisions that need to be made in order to correctly resolve the type of
a value. It also means that operations on Option types needed to be
language-level rather than implemented in terms of the language itself.
For example, consider:
```
val arr: Int? = []
if arr |arr| {
  if arr[0] |item| {
    item + 1
  } else {
    None
  }
} else {
  None
}
```
Now that Option values are backed by a prelude-defined enum `Option<T>`,
the API can be expanded to include `flatMap`/`map` methods; using those
method, the above example becomes much simpler:
```
arr.flatMap(arr => arr[0].map(item => item + 1))
// or, to be less fancy
arr.flatMap(arr => if arr[0] |item| Some(item + 1) else None)
```
Arguably this is easier to understand and also results in fewer
mistakes; in the first example above the value of `None` was returned in
each case, but if there were additional default logic that should be
applied, it'd have to be duplicated in each else-block (or else the
conditional logic would have to be extracted up a layer so the default
logic can be applied). Using an `orElse` method on `Option<T>` this can
become
```
arr
  .flatMap(arr => arr[0].map(item => item + 1))
  .orElse(() => None)
```
Returning `None` of course is the default, but this just illustrates the
benefit of having a fluent API to describe operations on Option values
rather than having to fall back to clunky language-level syntax.

Anyway, this was a large refactoring effort across the board, and
resulted in some additional benefits and reduction in surface area as
well - chief among them is the enforcement of both then- and else-blocks
when an if is used as an expression. Previously, the value would have
been implicitly wrapped in an Option type (or rather, "nullable" to be
more precise). I dislike the idea of such hidden logic/transformation
taking place behind the scenes.