Transforming invocations of `Some` to `Option.Some`
When parsing, if we see an invocation of `Some`, transform it to be
`Option.Some` instead, to construct an instance of the enum. The enum of
course can still be constructed directly, but since the Option type is
such a cornerstone of the language, it makes sense to have
out-of-the-box support for constructing it in a non-verbose way. The
same treatment will be given to `None` as well in a future changeset (as
well as `Result.Ok` and `Result.Err`, but that'll come in the future).

There's also an explicit declaration of `func Some<T>` to handle the
case where its use subverts the parser transformation. In typical code,
this function will never be invoked (since the parser will transform
apparent invocations of it to be `Option.Some`), but it's of course
possible to pass it around as a function value.