Selfhost: match special cases
Add handling for special cases `Ok` and `Err`. Just like the reference
implementation, this is a parser-level transformation which results in
match cases `Ok` and `Err` getting replaced with `Result.Ok` and
`Result.Err`, respectively. This also properly handles generics in match
case types and destructured variables.

This is part of attempting to typecheck the selfhosted compiler, so
there are a handful of other things in here as well.