Selfhost: compiling
This is the basic setup for actually generating an executable binary
from a typechecked AST in the selfhosted compiler. I decided to try out
QBE as it's significantly smaller in scope than LLVM, and seeing as how
I have to create my own convenience wrapper around it anyway, I want to
minimize the amount of work required. Who knows, maybe this will turn
out to be a mistake! But in my playing around with it so far, it seems
pretty promising.

This is a super rough draft at the moment, and the implementation in
`qbe.abra` will probably change quite a bit as I decide upon the APIs
that I want/need in order to write the compiler. There's also lots of
things that I'm punting on for now, such as creating the `._abra`
directory and the output `._abra/_main.ssa` file; there's a wrapper
`abra` bash script which takes care of all of that setup, as well as
running the `qbe` compiler (followed by `clang`).