Still getting set up
Building out the `qbe` module to feel more like a natural builder, akin
to `inkwell` (the rust library I had used in the reference
implementation for LLVM codegen).

Also, for now in order to validate that things seem to be working, I
just blindly pass the last expression in a module into `putchar`. This
is working pretty nicely to get things going, and has allowed me to
figure out a way to do Int literals while also laying out the groundwork
for the `qbe` module. This will go away once `println` is implemented,
which is first priority now that the setup is done.