Still roughing things out
Rather than using `putchar`, use `printf` on the last expression in the
module, and call the expression's `toString` method for the value. This
means there's some manner of method/function generation happening, which
is great progress! Also, I was wrestling with QBE's aggregate type
system a bit, and ultimately decided (for now, we'll see about the
future...) to forego them in lieu of just using raw `l` (`long`) types
to represent everything that's a pointer type (which is all
non-primitives, of course). To compensate for the difficulty that this
will very much introduce during debugging of QBE output code, I began a
system of adding "metadata" onto functions to help denote the
signatures. I think I'd also like to use that feature to denote
temporaries' types as well.