Selfhost: working towards compiling array literals
In order to do this, I need to be able to compile the
`Array.withCapacity` static method, as well as the `Array#push` instance
method. I also need to have a compiler-native implementation of the
`Pointer.malloc` (aka `pointer_malloc`) intrinsic. This poses an
interesting problem of whether or not I want to have some mechanism of
"resolved generics" in this implementation (otherwise, memory
allocations of generic types would have to assume 8 bytes for all
datatypes). The whole "resolved generics" thing, coupled with the
depth-first code generation approach (which basically implicitly
provides dead-code elimination), was the source of a ton of headaches in
the reference implementation. Without a debugger aside from `println`
statements, I worry that this would be even more of a headache here.