Selfhost: compiling (#444)
* Selfhost: compiling

This is the basic setup for actually generating an executable binary
from a typechecked AST in the selfhosted compiler. I decided to try out
QBE as it's significantly smaller in scope than LLVM, and seeing as how
I have to create my own convenience wrapper around it anyway, I want to
minimize the amount of work required. Who knows, maybe this will turn
out to be a mistake! But in my playing around with it so far, it seems
pretty promising.

This is a super rough draft at the moment, and the implementation in
`qbe.abra` will probably change quite a bit as I decide upon the APIs
that I want/need in order to write the compiler. There's also lots of
things that I'm punting on for now, such as creating the `._abra`
directory and the output `._abra/_main.ssa` file; there's a wrapper
`abra` bash script which takes care of all of that setup, as well as
running the `qbe` compiler (followed by `clang`).

* Still getting set up

Building out the `qbe` module to feel more like a natural builder, akin
to `inkwell` (the rust library I had used in the reference
implementation for LLVM codegen).

Also, for now in order to validate that things seem to be working, I
just blindly pass the last expression in a module into `putchar`. This
is working pretty nicely to get things going, and has allowed me to
figure out a way to do Int literals while also laying out the groundwork
for the `qbe` module. This will go away once `println` is implemented,
which is first priority now that the setup is done.

* Selfhost: compiling String literals

This involves compiling the String struct definition, as well as the
initializer function. I'm well on my way towards establishing the
foundations for this system. I'm liking QBE so far, although its error
messages are extremely sparse (this is by design, I suppose), and
documentation is sort of lacking (I spent way too long trying to figure
out how `malloc`'d Aggregate Types are supposed to work as return values).

* Still roughing things out

Rather than using `putchar`, use `printf` on the last expression in the
module, and call the expression's `toString` method for the value. This
means there's some manner of method/function generation happening, which
is great progress! Also, I was wrestling with QBE's aggregate type
system a bit, and ultimately decided (for now, we'll see about the
future...) to forego them in lieu of just using raw `l` (`long`) types
to represent everything that's a pointer type (which is all
non-primitives, of course). To compensate for the difficulty that this
will very much introduce during debugging of QBE output code, I began a
system of adding "metadata" onto functions to help denote the
signatures. I think I'd also like to use that feature to denote
temporaries' types as well.

* Adding some type-safety to API

I added an inkwell-inspired API for building instructions and managing
locals (more to come on that latter part). I'll continue to expand on
the idea as I think it's pretty fun.

* Adding toString for Int and Float

Also, add the ability to add comments to function declarations as well
as within blocks themselves.

* Add Bool#toString method

This also adds support for `jnz` and also introduces Labels.