Introduce `println` (and tests!)
This (likely) isn't going to be the final implementation of `println`
(but maybe it will be? I like how it's implemented as a compiler-native,
to avoid having to deal with `Any` yet. Not to mention the performance
benefits).
The main impetus with this is to be able to write tests for the
functionality I've built so far, so I know as I proceed if things break.
The idea is eventually to have 100% parity with the existing suite of
compiler tests, but for now the files are duplicated, with the majority
of the test cases are commented out.