Selfhost: working towards compiling array literals (#445)
* Selfhost: compiling

This is the basic setup for actually generating an executable binary
from a typechecked AST in the selfhosted compiler. I decided to try out
QBE as it's significantly smaller in scope than LLVM, and seeing as how
I have to create my own convenience wrapper around it anyway, I want to
minimize the amount of work required. Who knows, maybe this will turn
out to be a mistake! But in my playing around with it so far, it seems
pretty promising.

This is a super rough draft at the moment, and the implementation in
`qbe.abra` will probably change quite a bit as I decide upon the APIs
that I want/need in order to write the compiler. There's also lots of
things that I'm punting on for now, such as creating the `._abra`
directory and the output `._abra/_main.ssa` file; there's a wrapper
`abra` bash script which takes care of all of that setup, as well as
running the `qbe` compiler (followed by `clang`).

* Still getting set up

Building out the `qbe` module to feel more like a natural builder, akin
to `inkwell` (the rust library I had used in the reference
implementation for LLVM codegen).

Also, for now in order to validate that things seem to be working, I
just blindly pass the last expression in a module into `putchar`. This
is working pretty nicely to get things going, and has allowed me to
figure out a way to do Int literals while also laying out the groundwork
for the `qbe` module. This will go away once `println` is implemented,
which is first priority now that the setup is done.

* Selfhost: compiling String literals

This involves compiling the String struct definition, as well as the
initializer function. I'm well on my way towards establishing the
foundations for this system. I'm liking QBE so far, although its error
messages are extremely sparse (this is by design, I suppose), and
documentation is sort of lacking (I spent way too long trying to figure
out how `malloc`'d Aggregate Types are supposed to work as return values).

* Still roughing things out

Rather than using `putchar`, use `printf` on the last expression in the
module, and call the expression's `toString` method for the value. This
means there's some manner of method/function generation happening, which
is great progress! Also, I was wrestling with QBE's aggregate type
system a bit, and ultimately decided (for now, we'll see about the
future...) to forego them in lieu of just using raw `l` (`long`) types
to represent everything that's a pointer type (which is all
non-primitives, of course). To compensate for the difficulty that this
will very much introduce during debugging of QBE output code, I began a
system of adding "metadata" onto functions to help denote the
signatures. I think I'd also like to use that feature to denote
temporaries' types as well.

* Adding some type-safety to API

I added an inkwell-inspired API for building instructions and managing
locals (more to come on that latter part). I'll continue to expand on
the idea as I think it's pretty fun.

* Adding toString for Int and Float

Also, add the ability to add comments to function declarations as well
as within blocks themselves.

* Add Bool#toString method

This also adds support for `jnz` and also introduces Labels.

* Selfhost: working towards compiling array literals

In order to do this, I need to be able to compile the
`Array.withCapacity` static method, as well as the `Array#push` instance
method. I also need to have a compiler-native implementation of the
`Pointer.malloc` (aka `pointer_malloc`) intrinsic. This poses an
interesting problem of whether or not I want to have some mechanism of
"resolved generics" in this implementation (otherwise, memory
allocations of generic types would have to assume 8 bytes for all
datatypes). The whole "resolved generics" thing, coupled with the
depth-first code generation approach (which basically implicitly
provides dead-code elimination), was the source of a ton of headaches in
the reference implementation. Without a debugger aside from `println`
statements, I worry that this would be even more of a headache here.

* Fixing tests

* Field dot-accessor expressions

In order to compile `Array#push`, we need to be able to do field
accessors.

* Introduce `println` (and tests!)

This (likely) isn't going to be the final implementation of `println`
(but maybe it will be? I like how it's implemented as a compiler-native,
to avoid having to deal with `Any` yet. Not to mention the performance
benefits).
The main impetus with this is to be able to write tests for the
functionality I've built so far, so I know as I proceed if things break.
The idea is eventually to have 100% parity with the existing suite of
compiler tests, but for now the files are duplicated, with the majority
of the test cases are commented out.

* Updating some currently-failing tests

Because of limitations in the selfhosted compiler, compiling various
aspects of `prelude.abra` are not yet implemented. This provides a
really tricky setup where old tests will be incorrect until I can
update the selfhosted compiler to handle these cases (namely, in this
particular case, Arrays' capacity always starts at 16 rather than being
dependent on the number of items passed).

* Adding qbe to the build process

* fixing tar command

* add install/setup of qbe to github actions config (also caching)

* cleanup, and removing old ext/qbe stuff

* fixing gh actions config

* fixing gh actions config

* debugging gh actions config

* debugging gh actions config

* debugging gh actions config

* fixing yaml

* fixing yaml

* fixing yaml

* fixing yaml

* fixing yaml

* fixing yaml

* Uncomment prelude changes now that bindings are supported

* minor qol thing