Selfhost: getting `Array#toString` compiled
The forcing function here is to sucessfully compile the toString method
for an Array of Ints, and there's quite a lot that needs to happen in
order to accomplish that.
This changeset adds preliminary support for for-loops and enum variants,
as well as supporting condition bindings in if-exprs/stmts (as well as
supporting `Option` types in the condition, which is properly unwrapped
as a condition binding).