Array#toString is working!
In order to generate code properly for the Array#toString method, we need
to be able to call the appropriate T#toString method for the Array's inner
type. This inner type is always known at compile-time, so it's just a
matter of keeping track of the resolved generic values at each "layer"
as we DFS through the method calls. This is a bit more straightforward
than it was in the Rust reference implementation, but still fairly
complicated to get right (I'm sure I still have some lingering bugs
somewhere).
Also, the allowed names in QBE is pretty limited compared to LLVM, so I
need to add special logic when generating function/method names
unfortunately.
There's still outstanding work to get Float[] and Bool[] working, since
right now the Pointer intrinsics are still making certain assumptions
about the size (and type) of the values being stored and loaded.