Selfhost: getting `Array#toString` compiled (#446)
* Selfhost: getting `Array#toString` compiled

The forcing function here is to sucessfully compile the toString method
for an Array of Ints, and there's quite a lot that needs to happen in
order to accomplish that.
This changeset adds preliminary support for for-loops and enum variants,
as well as supporting condition bindings in if-exprs/stmts (as well as
supporting `Option` types in the condition, which is properly unwrapped
as a condition binding).

* Array#toString is working!

In order to generate code properly for the Array#toString method, we need
to be able to call the appropriate T#toString method for the Array's inner
type. This inner type is always known at compile-time, so it's just a
matter of keeping track of the resolved generic values at each "layer"
as we DFS through the method calls. This is a bit more straightforward
than it was in the Rust reference implementation, but still fairly
complicated to get right (I'm sure I still have some lingering bugs
somewhere).
Also, the allowed names in QBE is pretty limited compared to LLVM, so I
need to add special logic when generating function/method names
unfortunately.
There's still outstanding work to get Float[] and Bool[] working, since
right now the Pointer intrinsics are still making certain assumptions
about the size (and type) of the values being stored and loaded.

* getting toString working for Float[]

* Selfhost: Default-valued parameters

Remove the `rangeBy` function in the prelude and just use the `range`
function with a default-valued parameter. I had originally punted on
that but now that it's done I can revert those changes.