Fixing qbe output compilation (clang)
For some reason, the linker in clang on linux (for GH Actions) was
unable to find symbols from libgc.a despite it being listed in the
compilation targets alongside the output of qbe. If I flip the order, so
that libgc.a is _after_ the qbe output, it does work. This is opposite
to how it had worked with llvm, so I'm not quite sure what's going on
here. But it works now.