Selfhost/compiler/generated tostring (#450)
* Selfhost: compiling string concatenation

String concatenation is actually a bit more efficient in this
implementation versus the reference implementation, as it does not rely
on the `String#concat` method. Rather, it just calls the appropriate
`toString` method on either the lhs or rhs, and copies the two results
into a new String in-place. This may result in additional code
generation in situations where there's a lot of string concatenation
going on, but in the majority of cases this approach is more efficient
because it saves an  additional method call as well as avoids a variadic
function call (especially one which uses trait indirection to handle
`Any`).
In doing this, I also needed to ensure that the ResolvedGenerics layer
was applied prior to performing any intrinsic method calls (for both
instance methods and static methods), which resulted in an odd bug where
updating the allocated memory for a String's chars could overflow into
the memory region for the String instance itself and set the length
property accidentally. I don't really know why this was happening, as
enough memory should have been allocated (just using `malloc`) to fit
all of the characters, but it definitely was the issue. To "fix" this, I
pulled in libgc and switched to using `GC_malloc`/`GC_realloc` vs
standard `malloc`/`realloc` - this is something I needed to do anyway,
and it dodges the issue because the underlying strategies for memory
management differ. It's entirely possible that I'll encounter this issue
again, but now the implementation matches that of the reference
implementation, and I have yet to see this issue manifest there.

* ensuring compiler tests run in the correct working directory ('selfhost/')

* debugging gh actions

* debugging gh actions

* debugging gh actions

* Fixing qbe output compilation (clang)

For some reason, the linker in clang on linux (for GH Actions) was
unable to find symbols from libgc.a despite it being listed in the
compilation targets alongside the output of qbe. If I flip the order, so
that libgc.a is _after_ the qbe output, it does work. This is opposite
to how it had worked with llvm, so I'm not quite sure what's going on
here. But it works now.

* Selfhost: compiling generated toString methods