Selfhost: better PHIs
Phi instructions need to receive a list of pairs of labels/values, where
the label represents the label in the block where control flow had jumped
to arrive at the phi instruction. In the case of expressions which could
result in _many_ branches, we need to ensure that the _lastmost_ label
is the one that's passed into the list of phi pairs. This allows, among
other things, if-else expressions to work correctly.