Selfhost: compiling while loops (#455)
* Selfhost: compiling while loops

This also adds support for boolean condition bindings in if-exprs/stmts,
as well as in while loops (they're always set to true, of course, so
it's not particularly useful, but it's allowed for consistency). This
also unblocks many of the String method tests in the string test file,
as well as the majority of the tests in the array test file.

* Arrays as iterators in for-loops

When the subject of a for-loop is an array, call the `#iterator` method
on that array instance to obtain an iterator, and use the `#next` method
of that iterator instance in the for-loop.