Selfhost: compiling set/map literals
Compiling Sets requires the compilation of Maps, so may as well do them
both at once. Compiling Maps _also_ requires the compilation of Tuples,
which will be a slightly greater effort, and there was an issue that
needed to be fixed first regarding the way generics are resolved for
static functions. This fix ensures that the types tracked in
ResolvedGenerics have their type arguments recursively resolved using
the known resolved generics.

Additionally, this change includes a nice quality-of-life update by
removing the TypedIdentifierKind enum altogether - it's only ever the
'Named' variant now so it is not even necessary to differentiate.