Selfhost: adding tests for Set literals
I initially started out just trying to get the Set tests to work, but
ran into a very confusing issue (a segfault) when using the SetIterator
in the `Set#toString` method. This is due to an (admittedly sneaky)
issue with how a struct's field offset is calculated during accessor
expressions.