Selfhost: compiling set/map literals (#459)
* Selfhost: compiling set/map literals

Compiling Sets requires the compilation of Maps, so may as well do them
both at once. Compiling Maps _also_ requires the compilation of Tuples,
which will be a slightly greater effort, and there was an issue that
needed to be fixed first regarding the way generics are resolved for
static functions. This fix ensures that the types tracked in
ResolvedGenerics have their type arguments recursively resolved using
the known resolved generics.

Additionally, this change includes a nice quality-of-life update by
removing the TypedIdentifierKind enum altogether - it's only ever the
'Named' variant now so it is not even necessary to differentiate.

* Selfhost: toString, eq, hash for tuples

* Selfhost: fixing bug with hard-coded storel/loadl

* Selfhost: compiling map literals

There were quite a few smaller pieces needed in order to get the basic
initial tests to pass for maps, including temporarily removing the string
interpolation expression in `Map#toString` (I'll eventually revert that
change once string interpolation is implemented in the selfhosted compiler),
allowing Maps to be the subject of a for-loop, logic to correctly determine
the default param values mask when calling a function, and passing the
required arguments to the Map (and Set) `init` functions.

* Selfhost: adding tests for Set literals

I initially started out just trying to get the Set tests to work, but
ran into a very confusing issue (a segfault) when using the SetIterator
in the `Set#toString` method. This is due to an (admittedly sneaky)
issue with how a struct's field offset is calculated during accessor
expressions.

* Uncommenting set/map tests