Selfhost: compiling function value wrappers
Functions can be passed by value, and when they are it may be the case
that the target arity differs from the actual arity of the function
being passed. For example, if the expected function value has 2
parameters and the function being passed as a value only receives 1
parameter, this is allowed (the second parameter's value will simply by
thrown away when calling the function). To achieve this, we can create a
function wrapper to facilitate the discarding of the extraneous
parameters. There are other scenarios as well (including those which
account for functions with default-valued parameters) which will be
implemented in a future changeset.