Selfhost: compiling functions as values (#460)
* Selfhost: compiling functions as values

Expand upon the current `#buildCall` methods to accept an arbitrary
`Value` as opposed to a `QbeFunction` instance. This allows instances of
Functions (represented as structs, similar to tuples) to be passed
around as arbitrary values, and invoked using the stored function
pointer within.

* Selfhost: compiling function value wrappers

Functions can be passed by value, and when they are it may be the case
that the target arity differs from the actual arity of the function
being passed. For example, if the expected function value has 2
parameters and the function being passed as a value only receives 1
parameter, this is allowed (the second parameter's value will simply by
thrown away when calling the function). To achieve this, we can create a
function wrapper to facilitate the discarding of the extraneous
parameters. There are other scenarios as well (including those which
account for functions with default-valued parameters) which will be
implemented in a future changeset.

* Extract param-discarding function wrapper logic, and reuse it in the opt-param case

* Selfhost: compiling functions as values

Implementing the final case of arity mismatch between function being
passed as value and the required type. This also allowed for some minor
cleanup with the previous code (I'm sure there's also room for some
other improvements in that code as well...).

* Selfhost: compiling lambda expressions (non-closure)

* Selfhost: use lambdas in test files

* Fixing tests

* Selfhost: compiling closures for standalone functions

Note that this also only works at the moment when standalone functions
are called in the same scope as it's declared.

* Selfhost: compiling closure lambdas/functions-as-value

This provides an implementation for creating function values which
capture values from their declared environment. However, this still only
works for captures which are directly in the declared environment (no
nested closures yet).

* Selfhost: more closure compilation work

Ensuring closures defined within functions can close over parameters as
well as other variables defined within the function's scope.

* Selfhost: more closure compilation

Handle invocation of closure functions from within nested scopes.